add messages to ui-actions

+++ src/app/pages/author/author.page.ts
@@ -4,6 +4,7 @@ import Author from '../../core/models/author.model';
 import {UserService} from '../../core/services/user.service';
 import {AuthorService} from '../../core/services/author.service';
 import { Clipboard } from '@capacitor/clipboard';
+import {MessageService} from '../../core/services/message.service';
 
 @Component({
   selector: 'app-author',
@@ -18,6 +19,7 @@ export class AuthorPage implements OnInit, OnDestroy {
 
   constructor(private authorService: AuthorService,
               private userService: UserService,
+              private messageService: MessageService,
               private router: Router) {
   }
 
@@ -37,20 +39,38 @@ export class AuthorPage implements OnInit, OnDestroy {
 
   public confirmFriendRequest(friendId: string): void {
     this.authorService.addToFriendsList(friendId);
+    this.messageService.showToast('New friend added', 'light');
   }
 
-  public denyFriendRequest(friendId: string) {
-    this.authorService.removeFromPendingFriendsList(friendId);
+  public async denyFriendRequest(friendId: string): Promise<void> {
+    const confirmState = await this.messageService.showDialog(
+      'Are you sure to deny friend request?',
+      'middle',
+      'warning'
+    );
+
+    if (confirmState === 'confirm') {
+      this.authorService.removeFromPendingFriendsList(friendId);
+    }
   }
 
-  public removeFriend(friendId: string): void {
-    this.authorService.removeFromFriendList(friendId);
+  public async removeFriend(friendId: string): Promise<void> {
+    const confirmState = await this.messageService.showDialog(
+      'Are you sure to remove friend?',
+      'middle',
+      'warning'
+    );
+
+    if (confirmState === 'confirm') {
+      this.authorService.removeFromFriendList(friendId);
+    }
   }
 
   public async copyFriendLinkUrl() {
     await Clipboard.write({
       url: this.addFriendUrl
     });
+    this.messageService.showToast('Url copied!', 'light');
   }
 
   public logout() {
